Move env_settings.py content into es_utils.py
and add settings.ES_SETTINGS that an env can override with their own settings
